Improve `OneOfStrategy.do_draw` strategy drawing
